requirements.txt
cassandra-driver>=3.12
PyYAML==5.1.1
